refactor to use abstract parent struct
Having a generic parent struct provides a few benefits:

First, and likely most in the long term, it provides a common base type for use in third-party code that wants to accept any kind of changeset. And because the parent struct is generic, other code can also access the base type to provide more typesafe code just by taking a Changeset as an argument.

Second, it helps distinguish the public, documentable API of Changesets from implementation details. A lot of methods on Changesets are high-level combinators of lower level operations. These low level operations rely on macro-generated (or, at least, base-type-specific) code to provide type safety, but with that abstraction, the high level combinators can work without that knowledge. Splitting these into an "API parent" and "implementation children" makes this distinction more obvious.