validations and errors, `has_field?`
Validations allow for structured assertions on the effective data contained by a changeset (either in accepted changes or from the stored instance) to ensure that the final state meets some set of requirements. The first validation added is `validate_required`, which asserts the presence of values as anything other than `nil`.

Failed validations result in errors being added to the changeset, and the changeset being marked as invalid. Errors are stored as an array of `ChangesetError` objects and can be referenced through `changeset.errors`.

`has_field?` is used to validate that a field name given to a method represents a field that exists in the changeset. Generally, if this method returns false, the field is ignored from use for the remainder of the calling function.