JSON casting support.
`JSON::Any` _values_ are now handled properly by the typecaster. The implementation is made to be an example of how the typecasting system can be extended by end users to handle non-primitive types.

There's one annoyance in the flow in that `cast` relies on `has_key?`, which is not defined by `JSON::Any`, so passing params to a changeset requires calling `json.as_h`. I think this should be handled by extending `JSON::Any` with a `has_key?` method, rather than trying to mutate the Changeset's methods. This can also provide an example of making user-defined types compatible as parameters for changesets, which is an added bonus.