Don't allow casting _to_ nil, better bool handling, and more JSON.
Casting _to_ nil is separate from casting _from_ nil. The latter is necessary for handling nilability of fields, but the former is only valid when the incoming value is already nil. Even the string `"nil"` is not allowed, as "nil" is not a universal representation of the value (for example, JSON uses "null", some might use empty strings, etc.).

For booleans, however, the strings "false" and "true" are allowed as the representations are more universal.

JSON handling is also improved to delegate the raw value to the native cast methods, meaning it can bypass all the special cases from native casting and be safe from future changes.