Overly-magical DSL :/
In an ideal world, Changeset models could be generated just off of default `property` field declarations, similar to how `JSON::Serializable` works, but there are a few reasons that doesn't work (at least for now):

- `DB.mapping` doesn't work that way yet. It still requires a named tuple argument to take in properties. Hopefully this changes in Crystal 0.31?
- `macro finished` doesn't have instance variable information :( This one is really sad, because it would make everything much easier, just `include Change`, and it generates the complete Changeset type, copying the instance variables as properties and so on. But `@type.instance_vars` is empty everywhere outside of method definitions, so it can't define new properties directly.

As a compromise, the `macro included` hook is able to define the Changeset struct with it's own constants (e.g., `FIELDS`), and then the `field` macro can add properties to it. This keeps the door open to adding `belongs_to` and other combo macros like `timestamps` down the road, while also keeping the including module clean of any new definitions.