add iterators for changes and fields.
Similar to `get_change` and `get_field`, `each_change` and `each_field` allow a caller to iterate the stored changes/fields on the changeset, but without requiring any prior knowledge about the changeset to access them.

Unlike the current behavior of `changes_hash`, the types of values are preserved when yielding.